setup/requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
boto3==1.24.4
botocore==1.27.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.5
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
kavenegar==1.1.2
paramiko==2.11.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
